add `PyAudioNode`
This class may be used as a base class to create new audio node
subclasses only on the Python side (no JS code required). This is useful
for implementing custom or high-level audio nodes on top of existing
nodes.

Although a `PyAudioNode` object behaves like an audio node, it is not
itself an `AudioNode`, i.e., it is not an ipywidgets widget. It may be
built on top of other audio nodes or params (input, output...), which
ultimately should be ipytone widgets in order to process some audio
signal.