add PyAudioNode (#60)
* add `PyAudioNode`

This class may be used as a base class to create new audio node
subclasses only on the Python side (no JS code required). This is useful
for implementing custom or high-level audio nodes on top of existing
nodes.

Although a `PyAudioNode` object behaves like an audio node, it is not
itself an `AudioNode`, i.e., it is not an ipywidgets widget. It may be
built on top of other audio nodes or params (input, output...), which
ultimately should be ipytone widgets in order to process some audio
signal.

* add channel properties

* lint js

* allow connect a node to a PyAudioNode

* fix unrelated bug with Gain

The init gain value was ignored, due to the name of the parameter in
Tone.js ('gain' != 'value').

* fix PyToneAudioNode input and output nodes

I/O node were always undefined before.

* misc. fixes

* clean-up + proper support of nested py audio nodes

The encapsulated internal node of a `PyAudioNode` has
always widgets (or None) as input and output nodes.

Graph connection will go through pure-python source/destination input
and output nodes to get actual widgets.

* add tests

* lint py

* more tests